LayoutTests/platform/chromium/fast/js/reserved-words-strict-expected.txt

This file checks which ECMAScript 3 keywords are treated as reserved words in strict mode.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


SHOULD BE RESERVED:
FAIL isReserved('implements') should be true. Was false.
FAIL isReserved('interface') should be true. Was false.
FAIL isReserved('let') should be true. Was false.
FAIL isReserved('package') should be true. Was false.
FAIL isReserved('private') should be true. Was false.
FAIL isReserved('protected') should be true. Was false.
FAIL isReserved('public') should be true. Was false.
FAIL isReserved('static') should be true. Was false.
FAIL isReserved('yield') should be true. Was false.

SHOULD NOT BE RESERVED:
PASS isReserved('abstract') is false
PASS isReserved('boolean') is false
PASS isReserved('byte') is false
PASS isReserved('char') is false
PASS isReserved('double') is false
PASS isReserved('final') is false
PASS isReserved('float') is false
PASS isReserved('goto') is false
PASS isReserved('int') is false
PASS isReserved('long') is false
PASS isReserved('native') is false
PASS isReserved('short') is false
PASS isReserved('synchronized') is false
PASS isReserved('throws') is false
PASS isReserved('transient') is false
PASS isReserved('volatile') is false

PASS successfullyParsed is true

TEST COMPLETE
